# storage/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "短網址服務 API 文檔",
        "description": "提供短網址生成與管理功能",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "本地開發環境"
        }
    ],
    "paths": {
        "/api/v1": {},
        "/api/v1/shorten": {
            "post": {
                "tags": [
                    "Short URLs"
                ],
                "summary": "建立短網址",
                "operationId": "e602533ae36e3921c17967b7cb0542e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "original_url"
                                ],
                                "properties": {
                                    "original_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com"
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-31 23:59:59"
                                    },
                                    "max_clicks": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "短網址建立成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "original_url": {
                                                    "type": "string"
                                                },
                                                "short_url": {
                                                    "type": "string"
                                                },
                                                "short_code": {
                                                    "type": "string"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "max_clicks": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證錯誤"
                    }
                }
            }
        },
        "/s/{code}": {
            "get": {
                "tags": [
                    "Short URLs"
                ],
                "summary": "重定向到原始網址",
                "description": "根據短網址代碼重定向到對應的原始網址",
                "operationId": "06aac599dfaca1872e45f24d996c3694",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "短網址代碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "301": {
                        "description": "重定向到原始網址"
                    },
                    "410": {
                        "description": "短網址已過期或已達到點擊上限"
                    },
                    "404": {
                        "description": "短網址不存在"
                    }
                }
            }
        },
        "/api/v1/stats/{code}": {
            "get": {
                "tags": [
                    "Short URLs"
                ],
                "summary": "獲取短網址統計信息",
                "operationId": "34ce5a2b3c121421b1ea3c403e022a1b",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "短網址代碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功獲取統計信息",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "original_url": {
                                                    "type": "string"
                                                },
                                                "short_url": {
                                                    "type": "string"
                                                },
                                                "short_code": {
                                                    "type": "string"
                                                },
                                                "total_clicks": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                },
                                                "stats": {
                                                    "properties": {
                                                        "total_clicks": {
                                                            "type": "integer"
                                                        },
                                                        "unique_visitors": {
                                                            "type": "integer"
                                                        },
                                                        "browsers": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "browser": {
                                                                        "type": "string"
                                                                    },
                                                                    "count": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "platforms": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "platform": {
                                                                        "type": "string"
                                                                    },
                                                                    "count": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "devices": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "device_type": {
                                                                        "type": "string"
                                                                    },
                                                                    "count": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "referrers": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "referer": {
                                                                        "type": "string"
                                                                    },
                                                                    "count": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "clicks_by_date": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "date": {
                                                                        "type": "string",
                                                                        "format": "date"
                                                                    },
                                                                    "count": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "短網址不存在"
                    }
                }
            }
        },
        "/api/v1/url/{code}": {
            "delete": {
                "tags": [
                    "Short URLs"
                ],
                "summary": "刪除短網址",
                "operationId": "42d03849d459ea4723a92e52789f92fb",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "短網址代碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "短網址刪除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "短網址已刪除"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授權"
                    },
                    "403": {
                        "description": "沒有權限"
                    },
                    "404": {
                        "description": "短網址不存在"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Short URLs",
            "description": "短網址相關操作"
        }
    ]
}